+++ src/main/java/de/maxhenkel/car/integration/IMC.java
@@ -1,6 +1,6 @@
 package de.maxhenkel.car.integration;
 
-//import de.maxhenkel.car.integration.theoneprobe.TheOneProbeModule;
+import de.maxhenkel.car.integration.theoneprobe.TheOneProbeModule;
 
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.InterModComms;
@@ -12,8 +12,7 @@ public class IMC {
     @SubscribeEvent
     public static void enqueueIMC(InterModEnqueueEvent event) {
         if (ModList.get().isLoaded("theoneprobe")) {
-            // TODO
-            // InterModComms.sendTo("theoneprobe", "getTheOneProbe", TheOneProbeModule::new);
+            InterModComms.sendTo("theoneprobe", "getTheOneProbe", TheOneProbeModule::new);
         }
     }