+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed potential crash
+- Re implemented The One Probe integration